online_store/requirements.txt
asgiref==3.5.2
dj-loguru==0.2.2
Django==4.1.1
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-loguru==1.0.3
django-prometheus==2.2.0
django-taggit==3.0.0
djangorestframework==3.14.0
gunicorn==20.1.0
loguru==0.6.0
Pillow==9.2.0
prometheus-client==0.15.0
psycopg2-binary==2.9.1
pystache==0.6.0
pytz==2022.4
sqlparse==0.4.2
static==1.1.1
